chore(tests): add cluster shuffle to determinism test and consolidate tests
